+++ R/bundle.R
@@ -526,13 +526,13 @@ inferRPackageDependencies <- function(appMode, hasParameters, documentsHavePytho
   }
   if (appMode == "quarto-shiny") {
     # Quarto Shiny documents are executed with rmarkdown::run
-    deps <- c(deps, "rmarkdown", "shiny")
+    deps <- c(deps, "rmarkdown", shinyDeps())
   }
   if (appMode == "rmd-shiny") {
-    deps <- c(deps, "rmarkdown", "shiny")
+    deps <- c(deps, "rmarkdown", shinyDeps())
   }
   if (appMode == "shiny") {
-    deps <- c(deps, "shiny")
+    deps <- c(deps, shinyDeps())
   }
   if (appMode == "api") {
     deps <- c(deps, "plumber")
@@ -541,6 +541,15 @@ inferRPackageDependencies <- function(appMode, hasParameters, documentsHavePytho
     deps <- c(deps, "reticulate")
   }
   unique(deps)
+}
+
+# With shiny v1.7.2 or higher, renderPlot() defaults to ragg for png rendering
+# (if it's installed), so include it if it's installed locally (this way users
+# will get consistent PNG results without needing to add library(ragg) to the
+# app). https://github.com/rstudio/shiny/pull/3654
+shinyDeps <- function() {
+  include_ragg <- isAvailable("ragg") && isAvailable("shiny", "1.7.1.9000")
+  c("shiny", if (include_ragg) "ragg")
 }
 
 isWindows <- function() {